Added - added translation

+++ frontend/src/pages/PostPage.jsx
@@ -1,32 +1,26 @@
-import { Flex, Text, Image, Box, Divider, Button, Spinner, useDisclosure } from "@chakra-ui/react";
+import { Flex, Text, Image, Box, Divider, Button, Spinner } from "@chakra-ui/react";
 import { Avatar } from "@chakra-ui/avatar";
 import { useParams, useNavigate } from "react-router-dom";
+import { DeleteIcon } from "@chakra-ui/icons";
 import Actions from "../components/Actions";
-import { CommentWithActions } from "../components/Comment";
+import Comment from "../components/Comment";
 import useGetUserProfile from "../hooks/useGetUserProfile";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import useShowToast from "../hooks/useShowToast";
 import { formatDistanceToNow } from "date-fns";
+import { useRecoilValue, useRecoilState } from "recoil";
+import userAtom from "../atoms/userAtom";
 import postsAtom from "../atoms/postsAtom";
-import MenuActions from "../components/MenuActions";
-import { useRecoilState } from "recoil";
-import DeleteModal from "../components/DeleteModal";
-import { DELETE_MODAL_TYPES } from "../constants/deleteModal.constants";
+import { useTranslation } from "react-i18next";
 const PostPage = () => {
   const { isLoading, user } = useGetUserProfile();
   const [posts, setPosts] = useRecoilState(postsAtom);
   const showToast = useShowToast();
   const { pid } = useParams();
+  const currentUser = useRecoilValue(userAtom);
   const navigate = useNavigate();
-  const [isDeleting, setIsDeleting] = useState(false);
-  const { isOpen, onOpen, onClose } = useDisclosure();
+  const {t} = useTranslation();
   const currentPost = posts[0];
-
-  const goToPosterPage = (e) => {
-    e.preventDefault();
-    navigate(`/${user.username}`);
-  };
-
   useEffect(() => {
     const getPost = async () => {
       try {
@@ -49,7 +43,7 @@ const PostPage = () => {
   const handleDeletePost = async (e) => {
     try {
       e.preventDefault();
-      setIsDeleting(true);
+      if (!window.confirm("Are you sure to delete this post?")) return;
       const res = await fetch("/api/posts/" + currentPost._id, {
         method: "DELETE",
         headers: {
@@ -62,27 +56,11 @@ const PostPage = () => {
         showToast("Error", data.error, "error");
         return;
       }
-      setPosts(posts.filter((p) => p._id !== currentPost._id));
       showToast("Success", "Post deleted", "success");
-      navigate("/");
+      navigate(`/${user?.username}`);
     } catch (error) {
       showToast("Error", error, "error");
-    } finally {
-      setIsDeleting(false);
     }
-  };
-
-  const handleCopyLink = (e) => {
-    e.preventDefault();
-    const link = `${window.location.origin}/${user.username}/post/${currentPost._id}`;
-    navigator.clipboard.writeText(link).then(
-      () => {
-        showToast("Success", "Link copied!", "success");
-      },
-      (err) => {
-        showToast("Error", err, "error");
-      },
-    );
   };
 
   if (!user && isLoading) {
@@ -101,44 +79,26 @@ const PostPage = () => {
     <>
       <Flex>
         <Flex w={"full"} alignItems={"center"} gap={3}>
-          <Avatar
-            src={user?.profilePic}
-            size={"md"}
-            name={user?.username}
-            cursor={"pointer"}
-            onClick={goToPosterPage}
-          />
-          <Flex gap={1} alignItems={"center"}>
-            <Text
-              fontSize={"sm"}
-              fontWeight={"bold"}
-              cursor={"pointer"}
-              onClick={goToPosterPage}
-              _hover={{ textDecor: "underline" }}
-            >
+          <Avatar src={user?.profilePic} size={"md"} name={user?.username} />
+          <Flex>
+            <Text fontSize={"sm"} fontWeight={"bold"}>
               {user?.username}
             </Text>
-            <Image src="/verified.png" w={4} height={4}></Image>
-            <Text fontSize={"xs"} width={"auto"} textAlign={"left"} color={"gray.light"}>
-              {formatDistanceToNow(new Date(currentPost.createdAt))} ago
-            </Text>
+            <Image src="/verified.png" w={4} height={4} ml={3}></Image>
           </Flex>
         </Flex>
-        <MenuActions poster={user} onCopyLink={handleCopyLink} onDelete={onOpen} />
-        <DeleteModal
-          isOpen={isOpen}
-          onClose={onClose}
-          onDelete={handleDeletePost}
-          isLoading={isDeleting}
-          type={DELETE_MODAL_TYPES.post}
-        />
+        <Flex gap={4} alignItems={"center"}>
+          <Text fontSize={"xs"} width={48} textAlign={"right"} color={"gray.light"}>
+            {formatDistanceToNow(new Date(currentPost?.createdAt))} ago
+          </Text>
+
+          {currentUser?._id === user?._id && <DeleteIcon size={20} onClick={handleDeletePost} cursor={"pointer"} />}
+        </Flex>
       </Flex>
       <Text my={3}>{currentPost.text}</Text>
-      {currentPost.img && (
-        <Box borderRadius={6} overflow={"hidden"} border={"1px solid"} borderColor={"gray.light"}>
-          <Image src={currentPost?.img} w={"full"} />
-        </Box>
-      )}
+      <Box borderRadius={6} overflow={"hidden"} border={"1px solid"} borderColor={"gray.light"}>
+        <Image src={currentPost?.img} w={"full"} />
+      </Box>
       <Flex gap={3} my={3}>
         <Actions post={currentPost} />
       </Flex>
@@ -148,14 +108,14 @@ const PostPage = () => {
         <Flex>
           <Flex gap={2} alignItems={"center"}>
             <Text fontSize={"2xl"}>👋</Text>
-            <Text color={"gray.light"}>Get the app to like, reply and post.</Text>
+            <Text color={"gray.light"}>{t("getApp")}</Text>
           </Flex>
         </Flex>
-        <Button>Get</Button>
+        <Button>{t("get")}</Button>
       </Flex>
       <Divider my={4} />
       {currentPost.replies.map((reply) => (
-        <CommentWithActions
+        <Comment
           key={reply._id}
           reply={reply}
           lastReply={reply._id === currentPost.replies[currentPost.replies.length - 1]._id}
